+++ crates/cli/tests/install_uninstall_test.rs
@@ -80,6 +80,35 @@ mod install_uninstall {
     }
 
     #[test]
+    fn installs_via_prototools() {
+        let sandbox = create_empty_proto_sandbox();
+        sandbox.create_file(".prototools", "node = \"17\"");
+
+        sandbox
+            .run_bin(|cmd| {
+                cmd.arg("install").arg("node");
+            })
+            .success();
+
+        assert!(sandbox.path().join(".proto/tools/node/17.9.1").exists());
+    }
+
+    #[test]
+    fn installs_latest_if_no_version() {
+        let sandbox = create_empty_proto_sandbox();
+
+        assert!(!sandbox.path().join(".proto/tools/node").exists());
+
+        sandbox
+            .run_bin(|cmd| {
+                cmd.arg("install").arg("node");
+            })
+            .success();
+
+        assert!(sandbox.path().join(".proto/tools/node").exists());
+    }
+
+    #[test]
     fn installs_and_uninstalls_proto() {
         let sandbox = create_empty_proto_sandbox();
         let tool_dir = sandbox.path().join(".proto/tools/proto/0.45.0");